FileReader Complete
The FileLoader and FileWriter have been completed.
A host of ConfigActors have been created for the save of ensuring we can maintain our current config details between runs, as well as to help with the save/load process.
Interfaced with the GUI to provide standard save/load windows seen in most applications.

Tested and ensured that the load/save both produce the same file exactly.